requirements.txt
PyQt5>=5.15.0
mutagen>=1.45.0
Pillow>=9.0.0
yt-dlp>=2023.0.0
ytmusicapi>=0.25.0
customtkinter>=5.0.0
